1.21 GIGAWATTS THAT'S FAST, AND F-IN TRACERS
did some refactoring to support many of the changes. Changed the way rendering is done so it renders from the middle out horizontally; rendering uses a loop now rather than a queue, as well. Added tracer lines on top of the fractal that follow the event horizon of the drawing so you can visually see what part of the fractal is being updated. Changed event handling to allow zooming out with the right mouse button. Added the ability to shift the colors without recomputing the image, so you can adjust the colors even after it's been generated.