AH THAT CONTROL FEELS GOOD
changed how the zoom window is calculated so zooming is a bit more natural. Now the result of clicking is that the point that was clicked will remain in the same spot on the canvas -- essentially the zoom is done around the clicked spot but without centered on that spot. To center on a click you can hold shift and this won't zoom. Also fixed a small thing in the page button javascript that broke in a previous refactor